Refactored so each api is in its own file
