+++ backend/ht_buses_app/views/routes/route_delete.py
