+++ backend/ht_buses_app/views/schools/school_delete.py
