+++ backend/ht_buses_app/views/schools/schools_view.py
