+++ backend/ht_buses_app/views/students/student_delete.py
